bfloat16

+++ rlhf_ilql/train.py
@@ -104,7 +104,7 @@ if __name__ == '__main__':
     pl_model = MyILQLTransformer(config=config,model_args=model_args,training_args=training_args,lora_args=lora_args,ilql_args=ilql_args,
                                  load_in_8bit=load_in_8bit,device_map={"": trainer.fabric.local_rank} if trainer.world_size > 1 else "auto")
 
-  
+    pl_model.bfloat16()
     ckpt_path = './best_ckpt/best.pt'
     if not data_args.convert_onnx:
         #  只恢复权重 ， 不恢复步数和优化器 ，
